third_party/blink/web_tests/accessibility/selection-change-notification-on-selection-removed.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="textbox" role="textbox" contenteditable autofocus>
  <p>
    One paragraph in an ARIA textbox.
  </p>
</div>

<script>
'use strict';

let is_first_selection_event = true;

function axElementById(id) {
    return accessibilityController.accessibleElementById(id);
}

async_test((t) => {
  document.addEventListener('DOMContentLoaded', () => {
    const textbox = document.getElementById('textbox');
    // Set up initial selection, which will be cleared later.
    const selectionRange = document.createRange();
    selectionRange.selectNodeContents(textbox);
    getSelection().removeAllRanges();
    getSelection().addRange(selectionRange);

    const axRoot = accessibilityController.rootElement;
    axRoot.setNotificationListener(t.step_func((notification, intents) => {
      if (notification == 'DocumentSelectionChanged') {
        if (is_first_selection_event) {
          is_first_selection_event = false;
          assert_array_equals(intents,
                              ['AXEventIntent(setSelection,none,character,forward)']);
          const axTextBox = axElementById('textbox');
          assert_equals(axRoot.selectionAnchorObject, axTextBox, 'selectionAnchorObject');
          assert_equals(axRoot.selectionAnchorOffset, 0, 'selectionAnchorOffset');
          assert_equals(axRoot.selectionFocusObject, axTextBox, 'selectionFocusObject');
          assert_equals(axRoot.selectionFocusOffset, 1, 'selectionFocusOffset');
          // Clear the selection.
          getSelection().removeAllRanges();
        } else {
          assert_array_equals(intents,
                              ['AXEventIntent(clearSelection,none,none,none)']);
          axRoot.unsetNotificationListener();
          assert_equals(axRoot.selectionAnchorObject, null, 'selectionAnchorObject');
          assert_equals(axRoot.selectionAnchorOffset, -1, 'selectionAnchorOffset');
          assert_equals(axRoot.selectionFocusObject, null, 'selectionFocusObject');
          assert_equals(axRoot.selectionFocusOffset, -1, 'selectionFocusOffset');
          t.done();
        }
      }
    }));
  });
});

</script>
